dry run mode for archive cleanup

+++ pghoard/archive_cleanup.py
@@ -6,7 +6,7 @@ See LICENSE for details
 """
 from pghoard import common, config, logutil, version
 from pghoard.rohmu import get_transfer
-from pghoard.rohmu.errors import InvalidConfigurationError
+from pghoard.rohmu.errors import InvalidConfigurationError, FileNotFoundFromStorageError
 import argparse
 import logging
 import os
@@ -28,18 +28,26 @@ class ArchiveCleanup:
         storage_config = common.get_object_storage_config(self.config, self.site)
         self.storage = get_transfer(storage_config)
 
-    def archive_cleanup(self):
+    def archive_cleanup(self, dry_run):
         basebackup_path = os.path.join(self.backup_site["prefix"], "basebackup")
         xlog_path = os.path.join(self.backup_site["prefix"], "xlog")
         basebackups = self.storage.list_path(basebackup_path)
         first_required_wal = min(bb["metadata"]["start-wal-segment"] for bb in basebackups)
         self.log.info("First required WAL segment is %r", first_required_wal)
+        total_bytes = 0
         for object_info in self.storage.list_iter(xlog_path, with_metadata=False):
             object_name = object_info["name"]
             segment = os.path.basename(object_name)
             if segment < first_required_wal:
                 self.log.info("Orphan WAL segment %r needs to be deleted", segment)
-                self.storage.delete_key(object_name)
+                if "size" in object_info:
+                    total_bytes += int(object_info["size"])
+                if not dry_run:
+                    try:
+                        self.storage.delete_key(object_name)
+                    except FileNotFoundFromStorageError:
+                        self.log.error('Storage report segment %r is not available', segment)
+        self.log.info("Total orphan WAL segments size is %s bytes", total_bytes)
 
     def run(self, args=None):
         parser = argparse.ArgumentParser()
@@ -47,6 +55,13 @@ class ArchiveCleanup:
                             version=version.__version__)
         parser.add_argument("--site", help="pghoard site", required=False)
         parser.add_argument("--config", help="pghoard config file", default=os.environ.get("PGHOARD_CONFIG"))
+        parser.add_argument(
+            "--dry-run",
+            help="only list redundant segments and calculate total file size but do not delete",
+            required=False,
+            default=False,
+            action="store_true"
+        )
         args = parser.parse_args(args)
 
         if not args.config:
@@ -54,7 +69,7 @@ class ArchiveCleanup:
             return 1
 
         self.set_config(args.config, args.site)
-        return self.archive_cleanup()
+        return self.archive_cleanup(args.dry_run)
 
 
 def main():
